move compiler creator in one function
