--- packages/paperclip-compiler-html/index.js
@@ -1,3 +1,0 @@
-exports.compile = () => {
-  return {};
-};
