+++ packages/paperclip-compiler-html/package.json
@@ -2,7 +2,7 @@
   "name": "paperclip-compiler-html",
   "version": "18.2.16",
   "description": "HTML compiler for Paperclip",
-  "main": "index.js",
+  "main": "lib/index.js",
   "scripts": {
     "lint": "eslint src",
     "test": "mocha lib/test",
