+++ packages/paperclip-compiler-html/src/test/basic-test.ts
@@ -1,6 +1,6 @@
 import * as React from "react";
 import { expect } from "chai";
-import { compile } from "../code-compiler";
+import { compile } from "..";
 import {
   TEST_SUITE,
   compileModules
@@ -9,9 +9,13 @@ import {
 describe(__filename + "#", () => {
   TEST_SUITE.forEach(([title, graph, contexts, config, expected]: any) => {
     it(title, async () => {
-      const modules = await compileModules(compile, {
-        react: React
-      })(graph, config);
+      const modules = await compileModules(
+        compile,
+        {
+          react: React
+        },
+        "mjs"
+      )(graph, config);
 
       const entry = modules["/entry.pc"]();
