--- packages/paperclip-compiler-react/src/code-compiler.ts
@@ -1,6 +1,0 @@
-import { compiler } from "paperclip-compiler-base-jsx";
-
-export const compile = compiler({
-  imports: `import React from "react";\n`,
-  vendorName: "React"
-});
