+++ packages/paperclip-compiler-react/src/test/basic-test.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 import { expect } from "chai";
-import { compile } from "../code-compiler";
+import { compile } from "..";
 import * as ez from "enzyme";
 
 import * as Adapter from "enzyme-adapter-react-16";
