+++ packages/paperclip-coverage/paperclip.config.json
@@ -1,0 +1,6 @@
+{
+  "compilerOptions": {
+    "outDir": "lib"
+  },
+  "srcDir": "./src"
+}
